Move configuration details for runtime module.
